LayoutTests/editing/selection/ios/select-word-for-replacement-extends-grammar-marker-expected.txt

Verifies that selectWordForReplacement extends to a multi-word grammar marker. This test requires WebKitTestRunner.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.hasGrammarMarker(3, 9) is true
PASS getSelection().toString() is "are going"
PASS successfullyParsed is true

TEST COMPLETE
